Nova-Insights-Backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
upstash-redis==1.0.0
gunicorn==21.2.0
reportlab==4.0.8
beautifulsoup4==4.12.2
